# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dispatch_internal_modules_incidents_application.CreateIncidentRequest:
    properties:
      caller_name:
        type: string
      caller_phone:
        type: string
      created_by_user_id:
        type: string
      description:
        type: string
      district_id:
        type: string
      facility_id:
        type: string
      incident_type_id:
        type: string
      landmark:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      parish:
        type: string
      patient_age_group:
        type: string
      patient_name:
        type: string
      patient_phone:
        type: string
      patient_sex:
        type: string
      priority_level_id:
        type: string
      questionnaire_code:
        type: string
      severity_level_id:
        type: string
      source_channel:
        type: string
      subcounty:
        type: string
      summary:
        type: string
      triage_notes:
        type: string
      triage_responses:
        items:
          $ref: '#/definitions/dispatch_internal_modules_incidents_application.TriageResponseInput'
        type: array
      village:
        type: string
    required:
    - incident_type_id
    - source_channel
    - summary
    type: object
  dispatch_internal_modules_incidents_application.TriageResponseInput:
    properties:
      question_code:
        type: string
      response_value:
        type: string
    required:
    - question_code
    - response_value
    type: object
  dispatch_internal_modules_incidents_application.UpdateIncidentRequest:
    properties:
      caller_name:
        type: string
      caller_phone:
        type: string
      description:
        type: string
      district_id:
        type: string
      facility_id:
        type: string
      incident_type_id:
        type: string
      landmark:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      notes:
        type: string
      parish:
        type: string
      patient_age_group:
        type: string
      patient_name:
        type: string
      patient_phone:
        type: string
      patient_sex:
        enum:
        - MALE
        - FEMALE
        - OTHER
        - UNKNOWN
        type: string
      priority_level_id:
        type: string
      questionnaire_code:
        type: string
      reported_at:
        type: string
      severity_level_id:
        type: string
      source_channel:
        enum:
        - SMS
        - USSD
        - CALL
        - MOBILE_APP
        - WEB_PORTAL
        - FACILITY_REFERRAL
        type: string
      status:
        enum:
        - NEW
        - PENDING_VERIFICATION
        - VERIFIED
        - AWAITING_ASSIGNMENT
        - ASSIGNED
        - ENROUTE
        - AT_SCENE
        - TRANSPORTING
        - COMPLETED
        - CANCELLED
        - ESCALATED
        - REJECTED
        type: string
      subcounty:
        type: string
      summary:
        type: string
      triage_notes:
        type: string
      triage_responses:
        items:
          $ref: '#/definitions/dispatch_internal_modules_incidents_application.TriageResponseInput'
        type: array
      verification_status:
        enum:
        - PENDING
        - VERIFIED
        - REJECTED
        type: string
      village:
        type: string
    type: object
  dispatch_internal_modules_incidents_application.UpdateIncidentStatusRequest:
    properties:
      notes:
        type: string
      status:
        type: string
    required:
    - status
    type: object
...
info:
  contact: {}
  description: Go modular monolith backend for emergency alert intake, ambulance dispatch,
    user management, and fleet readiness.
  title: Dispatch Backend API
  version: "1.0"
paths:
...
  /incidents:
    get:
      description: Returns paginated incidents
      parameters:
      - description: Incident status
        in: query
        name: status
        type: string
      - description: District ID
        in: query
        name: district_id
        type: string
      - description: Facility ID
        in: query
        name: facility_id
        type: string
      - description: Priority level ID
        in: query
        name: priority_id
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by incident number, summary, or patient name
        in: query
        name: search
        type: string
      - description: Sort field
        enum:
        - reported_at
        - created_at
        - status
        in: query
        name: sort_by
        type: string
      - description: Sort order
        enum:
        - ASC
        - DESC
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List incidents
      tags:
      - Incidents
    post:
      consumes:
      - application/json
      description: Creates an incident and optionally persists triage responses on
        creation
      parameters:
      - description: Incident payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dispatch_internal_modules_incidents_application.CreateIncidentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create incident with triage
      tags:
      - Incidents
  /incidents/{id}:
    get:
      description: Returns an incident by ID
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get incident by ID
      tags:
      - Incidents
    put:
      consumes:
      - application/json
      description: Updates incident attributes
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      - description: Incident update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dispatch_internal_modules_incidents_application.UpdateIncidentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update incident
      tags:
      - Incidents
  /incidents/{id}/status:
    patch:
      consumes:
      - application/json
      description: Updates incident lifecycle status
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      - description: Incident status payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dispatch_internal_modules_incidents_application.UpdateIncidentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update incident status
      tags:
      - Incidents
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"